requirements.txt
Django==1.4
argparse==1.2.1
distribute==0.6.24
django-appconf==0.5
django-compressor==1.1.2
django-extensions==0.8
django-haystack==1.2.7
django-registration==0.8
django-tastypie==0.9.11
ipython==0.12.1
mimeparse==0.1.3
path.py==2.2.2
python-dateutil==1.5
wsgiref==0.1.2
yolk==0.4.3
